Change test mock source and sink to use column based format, also fix small bugs that occured from it (#689)
* Start on object serializer

* Fix C# object converter

* Add char, guid and hashset converters

* rename methods

* change test source to use column based data output

* Change test sink to also use column based data

* fix warning

* fix warnings and remove commented or unused code

+++ tests/FlowtideDotNet.AcceptanceTests/CastTests.cs
@@ -342,7 +342,7 @@ namespace FlowtideDotNet.AcceptanceTests
 
             await WaitForUpdate();
 
-            AssertCurrentDataEqual(Users.Select(u => new { val = u.Active ? (double)1 : (double)0 }));
+            AssertCurrentDataEqual(Users.Select(u => new { val = u.Active ? (decimal)1 : (decimal)0 }));
         }
 
         [Fact]
